add json file source type
